Movie rating game: Ground Work
Today i have started work on a new project. This application will be a game that my friends and i used to play over discord during lockdown. The idea is to choose 3 movies that you think their imdb rating when added together will be as close as possible to the target. The player with the closest score wins. I am trying to utilize as much of my recent learning to build this app from custom hooks to fetch the movies needed from the API to state management with the use reducer to keep all my state logic in one place. So far i have gotten a large part of this up and running however i still need to implement the functionality for getting , checking and storing the ratings of each chosen film and the score of each player.