Added: Rating fetching
The original API call i am using to pull the movies does not itself contain the imdbRating, instead i am having to use the imdbID that i am given back for each movie to then make a second request to the API for the actual rating which is now retrieved and added to the users score after each movie is added. This was more tricky than expected while attempting to get this work flawlessly with useReducer i went through several possible solutions before arriving at this one which i feel with the useEffect works really nicely.

EDIT: I came back to this later in the evening before commiting these changes and i have also now gotten the scoring system for each player entirely working. It was definately a head scratcher trying to get this to work in the parameters of my reducer function but i think its implemented nicely. I have also added some validation for the movies being added to ensure the user isnt adding multiple of the same movie to cheat the scores. This i feel could look cleaner but gets the job done. There is still a few more bits of validation to go in as well as the game end conditions and screen.