Functionality
Today and yesterday i have implemented the logic for the win condition and added an ending screen for this case. I have displayed all players and sorted them with the closest score first and a winning message.

Fixed: I also refactored the FormAddPlayer component as the way i had it originally had created a bug with the form not submiting properly, which is now fixed.

Fixed:  There was a bug where if the next player were to select the same movie as the last player for their first movie this would mess with the validation and cause the entire application to break.  It was caused by the fact that our useEffect was not running because it depends on the selectedId which wasnt changing if the next player added the same movie as the last movie added. I fixed this by resetting the selectedMovieId each time the player changes.

Fixed: There was a bug where once the last player chose their final movie the application would try to change to a player that doesnt exist. I fixed this by adding another check before executing the change player function to make sure there is a player we can switch to first.