Finishing Touches
So yesterday and this morning i have spent adding the finishing touches to the app.

Added: The ability to choose the player whos turn you want it to be. Thanks to the useReducer function and keeping track of the activePlayer via their ID i am able to easily switch player by adding a new case to the switch statement and adding an onClick to the player card. I have also ensured this onClick is not applied before the app starts.

Added: I used another custom hook together with the useRef hook to make the the input field focus when pressing enter and so the user doesnt need to click the box for each new user.

Added: I added a component for the isLoading and the error variables i was returning from the useMovies hook to conditionally render the Loader component when the movies are loading to display to the user that something is happening.

Added: i have also used the activePlayer state variable to conditionally apply an active class to the activePlayer so the user has a visual representation of whos turn it is.

Added: I added a play again button for ease of use so the user can either reset completely and enter new players or play again with the same players and maybe add some new ones in.

Added: General styling changes and touches to make the application look a little cleaner and nicer. I am not yet 100% happy with the design of the end screen yet particularly the way the movies and their ratings are displayed and i will make a final update to this when im clearer on how i want it to look.

Fixed:  I was having an issue with the API calls being bounced back from my hosted version of this application and the movies nor their ratings were not being fetched, this was due to API security and was a simple fix of updating the URL for my fetch requests to be https instead of http.